move(cosmo): move typing to private API
