+++ astropy/cosmology/_io/row.py
@@ -47,7 +47,7 @@ from .mapping import from_mapping
 if TYPE_CHECKING:
     from collections.abc import Mapping
 
-    from astropy.cosmology._typing import _CosmoT
+    from astropy.cosmology._src.typing import _CosmoT
     from astropy.table import Table
