+++ astropy/cosmology/core.py
@@ -31,7 +31,7 @@ if TYPE_CHECKING:
     from collections.abc import Mapping
     from typing import Self
 
-    from astropy.cosmology._typing import _CosmoT
+    from astropy.cosmology._src.typing import _CosmoT
     from astropy.cosmology.funcs.comparison import _FormatType
 
 # Originally authored by Andrew Becker (becker@astro.washington.edu),
